2023年5月14日 02:01

+++ PrefixMatcher.h
@@ -7,7 +7,7 @@
 class BinaryTrieNode {
  public:
   BinaryTrieNode* children[2];
-  int routerNumber = -1;
+  size_t routerNumber = -1;
 
   BinaryTrieNode();
 };
